Add graphql client, query, examples
