+++ lib/graphql/operations/mutations/mutation-example.dart
@@ -1,0 +1,9 @@
+const String addStar = r'''
+  mutation AddStar($starrableId: ID!) {
+    action: addStar(input: {starrableId: $starrableId}) {
+      starrable {
+        viewerHasStarred
+      }
+    }
+  }
+''';
