+++ lib/graphql/operations/queries/query-example.dart
@@ -1,0 +1,23 @@
+const String readRepositories = r'''
+  query ReadRepositories($nRepositories: Int!) {
+    viewer {
+      repositories(last: $nRepositories) {
+        nodes {
+          __typename
+          id
+          name
+          viewerHasStarred
+        }
+      }
+    }
+  }
+''';
+
+const String testSubscription = r'''
+		subscription test {
+	    deviceChanged(id: 2) {
+        id
+        name
+      }
+		}
+''';
