+++ pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
     sdk: flutter
 
   graphql: ^3.0.0
+  union: ^0.0.3+1
   # OAuth
   http: ^0.12.1
   flutter_webview_plugin: ^0.3.11
